transfer shoppingList dummy array into seperate file
